+++ Conductor/code/StatusEffects/StatusEffectSteelguardState.cs
@@ -7,9 +7,9 @@ using System.Text;
 
 namespace Conductor.StatusEffects
 {
-    public class StatusEffectSteelguardState : StatusEffectState
+    public class StatusEffectSteelguardState : StatusEffectState 
     {
-        public override bool TestTrigger(InputTriggerParams inputTriggerParams, OutputTriggerParams outputTriggerParams, ICoreGameManagers coreGameManagers)
+        public override bool TestTrigger(InputTriggerParams inputTriggerParams, OutputTriggerParams outputTriggerParams, ICoreGameManagers coreGameManagers) 
         {
             CharacterState character = inputTriggerParams.attacked;
             if (character == null)
@@ -17,16 +17,21 @@ namespace Conductor.StatusEffects
             if (DamageHelper.IsPiercingDamage(inputTriggerParams.damageSourceCard, inputTriggerParams.damageSourceRelic, coreGameManagers, inputTriggerParams.attacker, inputTriggerParams.damageType))
                 return false;
             int armor = character.GetStatusEffectStacks("armor");
-            if (inputTriggerParams.damage > GetParamInt() && armor >= GetParamInt())
+            if (inputTriggerParams.damage > GetParamInt() && armor >= GetParamInt()) 
             {
                 outputTriggerParams.damage = GetParamInt();
+                outputTriggerParams.damageBlocked = inputTriggerParams.damage - outputTriggerParams.damage;
                 return true;
             }
             return false;
         }
-        public override int GetTriggerOrder()
+        public override int GetTriggerOrder() 
         {
             return -1;
+        }
+        public override int GetEffectMagnitude(int stacks = 1)
+        {
+            return GetParamInt();
         }
     }
 }